fix: review findings — all-empty concat schema, week components, Total Encounters label; move aggregation models into _patrol_summary
- concat_dataframes: all-empty (but unskipped) inputs now concat to keep
  the union of input columns instead of returning a bare frame
- decompose_datetime: 'week'/'weekofyear' via isocalendar (Series.dt.week
  was removed in pandas 2.0)
- TotalEncounterMetric: 'Total Encounters' title and count-mode label, so
  it no longer collides with TotalEventsMetric's 'Total Events' column
- extract encounter_metrics_to_summary_params for reuse by the trend
  chart config task
- move CountAggregation/SumOfColumnAggregation/make_aggregation_json_schema_extra
  next to their only consumer in _patrol_summary; drop the unused
  AggregationAnnotation alias
